ingestion-api/app/requirements.txt
flask
pytest
psycopg2
python-dotenv